Issue #42: Moving HxTrigger and HxRefresh handling to preHandle to be compatible with @ResponseBody
Due to the nature of how ResponseBody is processed, a request is committed (therefore silently immutable) *before* an interceptor's postHandle is ever called. Therefore, these headers were being lost in this ResponseBody situation (and maybe other unknowns). The fix is to set the headers instead in preHandle(), as it makes no difference to existing functionality but fixes this issue.